Add non class parser method in a separate file + Loop through headers
